refactored for testability
We extracted functionality into separate classes to be able to test the individual parts of the algorithm.
In this first step, only the DPLL parts were refactored.
Next step is to also extract the CDCL parts and unit test them.
We hopefully find a mistake that causes the bad CDCL performance.