Refactored for testibility, separated solver loops for CDCL and Poor Man's VSIDS
We also already fixed some bugs in the solver loop that prevented the propagation of the uips of learned clauses.
We extracted functionality into separate classes to be able to unit test them.
CDCL is still not working, but now we have clearer errors and the possibility to test individual parts of the algorithm as well as their integration.